+++ CHANGELOG.md
@@ -33,6 +33,8 @@
   - [#4325](https://github.com/wix/ricos/pull/4325) a11y close modals on esc click
 - `collapsible-list`
   - [#4317](https://github.com/wix/ricos/pull/4317) fix settings functionality tiptap
+- `image`
+  - [#4333](https://github.com/wix/ricos/pull/4333) fix tiptap `wixImage`
 
 </details>